services/wiki/requirements.txt
# Wiki Service - Service-Specific Dependencies
#
# This file contains ONLY dependencies specific to the Wiki Service.
# Shared dependencies (Flask, SQLAlchemy, etc.) are in the root requirements.txt
#
# To install all dependencies:
#   1. From project root: pip install -r requirements.txt
#   2. From this directory: pip install -r requirements.txt
#
# Or install both:
#   pip install -r requirements.txt -r services/wiki/requirements.txt

# Wiki-specific dependencies
Flask-CORS==4.0.0
PyYAML==6.0.1
watchdog==3.0.0
requests==2.31.0
# Note: HTML/Markdown conversion uses JavaScript libraries via subprocess
# No Python packages needed for turndown/marked

# Note: psycopg2-binary, Flask, Flask-SQLAlchemy, Flask-Migrate, python-dotenv,
# pytest, pytest-cov are in the root requirements.txt
# Tests use SQLite in-memory for faster execution
